new changes

+++ DHLWebAPI/Models/TblCardStatus.cs
@@ -10,11 +10,10 @@ namespace DHLWebAPI.Models
             TblCards = new HashSet<TblCards>();
         }
 
-        public int IdCardStatus { get; set; }
+        public byte IdCardStatus { get; set; }
         public string CardStatus { get; set; }
         public int UpdateBy { get; set; }
         public DateTime UpdateDate { get; set; }
-
 
         public virtual TblUsers UpdateByNavigation { get; set; }
         public virtual ICollection<TblCards> TblCards { get; set; }
